src/web/requirements.txt
fastapi>=0.100.0
dotenv
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langgraph-sdk>=0.1.61
python-dotenv>=1.0.0
pymongo>=4.3.0
jinja2>=3.0.0 
yfinance>=0.2.58
python-multipart
# passlib[bcrypt]>=1.7.4 # Comment out the old line
passlib==1.7.4
bcrypt==3.2.0
jwt
python-jose[cryptography]>=3.3.0